requirements.txt
torchtext
Pillow
torchvision==0.5.0
torch==1.4.0
numpy==1.18.1
matplotlib==3.2.0
